[EI-249] [web] refactor: Update import paths for indicator board metadata view model service
